==> scripts/2_matching.Rmd <==
---
title: "2_matching"
output: html_document
date: "2024-04-22"
---

```{r}
library(MatchIt)
library(tidyverse)
```

```{r}
master <- read.csv(file="../data/master.csv") 
master$location <- 'boston'
master_birkbeck <- read.csv(file='../data/master_Birkbeck.csv')
master_birkbeck$location <- 'birkbeck'
master <- rbind(master, master_birkbeck)
master_single <- master %>%
  distinct(ResponseId, .keep_all = TRUE)

master_birkbecksingle <- master_birkbeck %>%
  distinct(ResponseId, .keep_all = TRUE)

master_bostonsingle <- master %>%
  distinct(ResponseId, .keep_all = TRUE)
```

#histograms
```{r}
#eBMRQ scores by sample
master_birkbecksingle %>% 
  ggplot(aes(x=eBMRQ_total)) +
  geom_histogram(binwidth=5) +
  theme_classic() +
  xlab("eBMRQ Score") +
  ylab("Count") +
  ggtitle("Birkbeck Sample") 

master_bostonsingle %>% 
  ggplot(aes(x=eBMRQ_total)) +
  geom_histogram(binwidth=5) +
  theme_classic() +
  xlab("eBMRQ Score") +
  ylab("Count") +
  ggtitle("Boston Sample") 

#PAS no audio histograms by sample

master_birkbecksingle %>% 
  ggplot(aes(x=PASS_noaud)) +
  geom_histogram(binwidth=3) +
  theme_classic() +
  xlab("PAS Score (no auditory items)") +
  ylab("Count") +
  ggtitle("Birkbeck Sample") 

master_bostonsingle %>% 
  ggplot(aes(x=PASS_noaud)) +
  geom_histogram(binwidth=3) +
  theme_classic() +
  xlab("PAS Score (no auditory items)") +
  ylab("Count") +
  ggtitle("Boston Sample") 

master_bostonsingle %>% 
  ggplot(aes(x=eBMRQ_total)) +
  geom_histogram(binwidth=5) +
  theme_classic() +
  xlab("eBMRQ Score") +
  ylab("Count") +
  ggtitle("Boston Sample")
```


```{r}
mean(master_single$PASS_noaud) + sd(master_single$PASS_noaud)
quantile(master_single$eBMRQ_total, probs = c(.1), na.rm=TRUE)

master %>% distinct(ResponseId, .keep_all = TRUE) %>%
  ggplot(aes(x=PASS_noaud, y=eBMRQ_total))+
  geom_point(alpha=0.65) +
  geom_hline(yintercept = 69.6, linetype="dotted", color="blue") +
  geom_vline(xintercept = 16.98947, linetype="dotted", color="red") +
  theme_classic() +
   theme(legend.position="none")+
  xlab("PAS Scores (Sound Items Removed)") +
  ylab("Total eBMRQ Scores") 
  #geom_point(data=df.match, aes(color=group)) 


general_hedonics <- 
  master_single %>%
  subset(PASS_noaud<16.98947)

musical_anhedonics <- subset(general_hedonics, eBMRQ_total< 69.6)
musical_anhedonics$group = "Anhedonic"

hedonics <- subset(general_hedonics, eBMRQ_total > 69.6)
hedonics$group = "Hedonic"

potential_matchedcontrols <- rbind(musical_anhedonics, hedonics)

potential_matchedcontrols$group = ifelse(potential_matchedcontrols$group == 'Anhedonic', 1, 0)
potential_matchedcontrols$group <- as.factor(potential_matchedcontrols$group)
matched <- MatchIt::matchit(group ~ Gold_perceptualabilities + PASS_noaud + Gold_musicaltraining, data=potential_matchedcontrols, method='nearest', ratio=1)
summary(matched)
df.match <- MatchIt::match.data(matched)[1:ncol(potential_matchedcontrols)]

df.match$group = ifelse(df.match$group == '0', "Matched Controls", "Musical Anhedonics")

controls <- subset(df.match, group=="Matched Controls")
anhedonics <- subset(df.match, group=="Musical Anhedonics")

write.csv(df.match, file="../data/anhedonics_matchedcontrols.csv")

df.matchlong <- master %>%
  subset(ResponseId %in% df.match$ResponseId)

df.matchlong <- df.match %>%
  select(ResponseId, group) %>%
  merge(df.matchlong, by= "ResponseId")

write.csv(df.match, file="../data/anhedonics_matchedcontrols.csv")
write.csv(df.matchlong, file="../data/anhedonics_matchedcontrolslong.csv")
```
