==> scripts/3_analysis.Rmd <==
---
title: "3_analysis"
output: html_document
date: "2024-04-22"
---

```{r}
library(tidyverse)
library(lme4)
library(lmerTest)
library(emmeans)
```

```{r}
master_boston <- read.csv(file='../data/master.csv')
master_boston$location = "boston"
master_birkbeck <- read.csv(file='../data/master_Birkbeck.csv')
master_birkbeck$location = "birkbeck"

master <- rbind(master_boston, master_birkbeck) %>%
  dplyr::select(ResponseId:PASS_noaud, value:location)

write.csv(master, '../data/master_cleaned.csv')


matched <- read.csv(file='../data/anhedonics_matchedcontrolslong.csv')

matched <- matched %>%
  dplyr::select(ResponseId:group, Q409:PASS_noaud, value:location)


write.csv(matched, '../data/matched_clean.csv')
```

#cleaning
```{r}
liking_master <- master %>%
  subset(rating=="liking")

liking_matched <- matched %>%
  subset(rating=="liking")

speech_master <- master %>%
  subset(rating=="speech")

speech_matched <- matched %>%
  subset(rating=="speech")

#setting contrasts correctly for analyses
speech_master$stim_type <- as.factor(speech_master$stim_type)
speech_master$pre_post <- as.factor(speech_master$pre_post)
contrasts(speech_master$stim_type) <- c(-0.5,0.5)
speech_master$pre_post <- factor(speech_master$pre_post, levels = c("pre", "post"))
contrasts(speech_master$pre_post) <- c(-0.5,0.5)
speech_master$Q409 <- factor(speech_master$Q409, levels = c("Laurel", "Yanny"))
contrasts(speech_master$Q409) <- c(-0.5,0.5)

#for s2s + liking across sample
liking_master$stim_type <- as.factor(liking_master$stim_type)
liking_master$stim <- as.factor(liking_master$stim)
liking_master$pre_post <- as.factor(liking_master$pre_post)
liking_master$pre_post <- factor(liking_master$pre_post, levels = c("pre", "post"))
contrasts(liking_master$stim_type) <- c(-0.5,0.5)
contrasts(liking_master$pre_post) <- c(-0.5,0.5)
liking_master$Q409 <- factor(liking_master$Q409, levels = c("Laurel", "Yanny"))
contrasts(liking_master$Q409) <- c(-0.5,0.5)

speech_matched$pre_post <- factor(speech_matched$pre_post, levels = c("pre", "post"))
speech_matched$stim_type <- as.factor(speech_matched$stim_type)
speech_matched$group <- as.factor(speech_matched$group)
contrasts(speech_matched$stim_type) <- c(-0.5,0.5)
contrasts(speech_matched$group) <- c(0.5,-0.5)
contrasts(speech_matched$pre_post) <- c(-0.5,0.5)

liking_matched$pre_post <- factor(liking_matched$pre_post, levels = c("pre", "post"))
liking_matched$stim_type <- as.factor(liking_matched$stim_type)
liking_matched$group <- as.factor(liking_matched$group)
contrasts(liking_matched$stim_type) <- c(-0.5,0.5)
contrasts(liking_matched$group) <- c(0.5,-0.5)
contrasts(liking_matched$pre_post) <- c(-0.5,0.5)

master$pre_post <- factor(master$pre_post, levels = c("pre", "post"))

masterprepost <- pivot_wider(master, names_from = pre_post, values_from = value)

masterprepost$change <- masterprepost$post - masterprepost$pre
```

#descriptive stats
```{r}
#musicality rating histograms by sample
master_birkbeck %>% 
  subset(rating=="speech") %>%
  ggplot(aes(x=value)) +
  geom_histogram(binwidth=1) +
  theme_classic() +
  xlab("Musicality Rating") +
  ylab("Count") +
  ggtitle("Birkbeck Sample") +
  facet_grid(~pre_post)

master_boston %>% 
  subset(rating=="speech") %>%
  ggplot(aes(x=value)) +
  geom_histogram(binwidth=1) +
  theme_classic() +
  xlab("Musicality Rating") +
  ylab("Count") +
  ggtitle("Boston Sample") +
  facet_grid(~pre_post)

#liking rating histograms by sample

master_birkbeck %>% 
  subset(rating=="liking") %>%
  ggplot(aes(x=value)) +
  geom_histogram(binwidth=1) +
  theme_classic() +
  xlab("Liking Rating") +
  ylab("Count") +
  ggtitle("Birkbeck Sample") +
  facet_grid(~pre_post)

master_boston %>% 
  subset(rating=="liking") %>%
  ggplot(aes(x=value)) +
  geom_histogram(binwidth=1) +
  theme_classic() +
  xlab("Liking Rating") +
  ylab("Count") +
  ggtitle("Boston Sample") +
  facet_grid(~pre_post)

#laurel v yanny
master_birkbeck %>% 
  distinct(ResponseId, .keep_all = T) %>%
  ggplot(aes(x=Q409)) +
  stat_count() +
  theme_classic() +
  xlab("")+
  ylab("Count") +
  geom_text(aes(label = after_stat(count)),
    stat = "count", position = position_fill(vjust = 100), color="white",size=10
  ) +
  theme(
    axis.title.x = element_text(family="Arial", size=20,face="bold"),
    axis.title.y = element_text(family="Arial",size=20,face="bold"),
    axis.text.x = element_text(family="Arial", size=14,face="bold"),
    axis.text.y = element_text(family="Arial",size=14,face="bold"),
    strip.text = element_text(family="Arial",size=16,color='black',face="bold")
  )
```

#S2S across entire sample
```{r}
#FIGURE 1A
#making stim_type labels
labels <- expand.grid(stim_type=c("Control", "Illusion"),
                      pre_post="pre",
                      value=4,
                      rating="Musicality"
  )

labels$value <- ifelse(labels$stim_type=="Control", 2.5, 3.125)
  
liking_musciality_all <-
  master %>%
  mutate(rating = recode(rating, "liking" = "Liking", "speech" = "Musicality")) %>%
  mutate(rating = factor(rating, levels = c("Musicality", "Liking"))) %>%
  mutate(stim_type = recode(stim_type, "illusion" = "Illusion", "control" = "Control")) %>%
  ggplot(aes(x=pre_post, y=value, color=stim_type)) + 
 #ggdist::geom_dotsinterval() +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'errorbar', width = 0.05, alpha=0.5) +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'pointrange', alpha=0.5) +
  theme_bw()+ 
  #geom_jitter(width=0.4,height=0.4,alpha=0.1) +
  #geom_line(alpha=0.05, color="black")+
  #geom_point()+
 theme(legend.position = "none")+
  ylab("Rating") +
  xlab("") +
  scale_x_discrete(labels = c("Pre", "Post")) +
  #coord_cartesian(ylim = c(2,5)) +
   theme(
    axis.title.x = element_text(family="Arial", size=20,face="bold"),
    axis.title.y = element_text(family="Arial",size=20,face="bold"),
    axis.text.x = element_text(family="Arial", size=14,face="bold"),
    axis.text.y = element_text(family="Arial",size=14,face="bold"),
    strip.text = element_text(family="Arial",size=16,color='black',face="bold")
  ) + 
  facet_wrap(~rating, scales="free_y") +
  ggrepel::geom_label_repel(data=labels, aes(label=stim_type),min.segment.length = Inf, size=3.5) +
  scale_color_manual(values=c("#0072B2", "#D55E00" )) +
  scale_fill_manual(values=c("#0072B2","#D55E00"))

ggsave(liking_musciality_all,file='../figures/liking_musciality_all.png', width=9, dpi=1000)


#FIGURE1B
#collapse across stims for each P in each condition for each rating type:
delta <- 
masterprepost %>%
  group_by(ResponseId, rating, stim_type) %>%
  dplyr::summarize(change = mean(change, na.rm=TRUE))

change_liking_musciality_all <-
delta %>%
  mutate(rating = recode(rating, "liking" = "Liking", "speech" = "Musicality")) %>%
  mutate(rating = factor(rating, levels = c("Musicality", "Liking"))) %>%
  mutate(stim_type = recode(stim_type, "illusion" = "Illusion", "control" = "Control")) %>%
  ggplot(aes(x=stim_type, y=change, color=stim_type)) + 
 #ggdist::geom_dotsinterval() +
  theme_bw()+ 
  geom_jitter(width=0.3,alpha=0.5) +
  geom_boxplot(size=0.5, varwidth=TRUE, outliers = FALSE, color="black", fill=NA)+
  geom_hline(color="red", linetype="dashed", yintercept = 0)+
  #stat_summary(fun.data = 'mean_cl_normal', geom = 'errorbar', width = 0.025, color="black") +
  #stat_summary(fun.data = 'mean_cl_normal', geom = 'pointrange',color="black", size=0.1) +
  #geom_point()+
 theme(legend.position = "none")+
  ylab("Change in Rating (Post - Pre)") +
  xlab("") +
  #scale_x_discrete(labels = c("Musicality", "Liking")) +
  #coord_cartesian(ylim = c(2,5)) +
   theme(
    axis.title.x = element_text(family="Arial", size=20,face="bold"),
    axis.title.y = element_text(family="Arial",size=20,face="bold"),
    axis.text.x = element_text(family="Arial", size=14,face="bold"),
    axis.text.y = element_text(family="Arial",size=14,face="bold"),
    strip.text = element_text(family="Arial",size=16,color='black',face="bold")
  )+
  facet_grid(~rating) +
  scale_color_manual(values=c("#0072B2", "#D55E00" )) +
  scale_fill_manual(values=c("#0072B2","#D55E00"))

ggsave(change_liking_musciality_all,file='../figures/change_liking_musciality_all.png', width=7, dpi=1000)


#FIGURE 2A
#updating labels
labels$Q409 <- "Laurel Hearers"
labels$value <- labels$value- 0.05

labels$value <- ifelse(labels$stim_type =="Control", 2.55, 3.2)
#LYspeech <- 
speech_master %>%
  mutate(Q409 = recode(Q409, "Laurel" = "Laurel Hearers", "Yanny" = "Yanny Hearers"))%>%
  mutate(stim_type = recode(stim_type, "illusion" = "Illusion", "control" = "Control")) %>%
  ggplot(aes(x=pre_post, y=value, color=stim_type)) + 
 #ggdist::geom_dotsinterval() +
  stat_summary(fun.data = 'mean_se', geom = 'errorbar', width = 0.05, alpha=0.5) +
  stat_summary(fun.data = 'mean_se', geom = 'pointrange', alpha=0.5) +
  theme_bw()+ 
  #geom_jitter(width=0.4,height=0.4,alpha=0.1) +
  #geom_line(alpha=0.05, color="black")+
  #geom_point()+
 theme(legend.position = "none")+
  ylab("Musicality Rating") +
  xlab("") +
theme(
    axis.title.x = element_text(family="Arial", size=20,face="bold"),
    axis.title.y = element_text(family="Arial",size=20,face="bold"),
    axis.text.x = element_text(family="Arial", size=14,face="bold"),
    axis.text.y = element_text(family="Arial",size=14,face="bold"),
    strip.text = element_text(family="Arial",size=16,color='black',face="bold")
  ) +
  coord_cartesian(ylim=c(2,5))+
  facet_grid(~Q409) +
  scale_x_discrete(labels = c("Pre", "Post")) + 
  ggrepel::geom_label_repel(data=labels, aes(label=stim_type),min.segment.length = Inf, size=3.5) +
  scale_color_manual(values=c("#0072B2", "#D55E00" )) +
  scale_fill_manual(values=c("#0072B2","#D55E00"))


ggsave(LYspeech,file='../figures/LYspeech.png', width=9, dpi=1000)

LYliking <- 
liking_master %>%
mutate(stim_type = recode(stim_type, "illusion" = "Illusion", "control" = "Control")) %>%
  mutate(Q409 = recode(Q409, "Laurel" = "Laurel Hearers", "Yanny" = "Yanny Hearers"))%>%
ggplot(aes(x=pre_post, y=value, color=stim_type)) + 
 #ggdist::geom_dotsinterval() +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'errorbar', width = 0.05, alpha=0.5) +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'pointrange', alpha=0.5) +
  theme_bw()+ 
  #geom_jitter(width=0.4,height=0.4,alpha=0.1) +
  #geom_line(alpha=0.05, color="black")+
  #geom_point()+
 theme(legend.position = "none")+
  ylab("Liking Rating") +
  xlab("") +
theme(
    axis.title.x = element_text(family="Arial", size=20,face="bold"),
    axis.title.y = element_text(family="Arial",size=20,face="bold"),
    axis.text.x = element_text(family="Arial", size=14,face="bold"),
    axis.text.y = element_text(family="Arial",size=14,face="bold"),
    strip.text = element_text(family="Arial",size=16,color='black',face="bold")
  ) +
  #coord_cartesian(ylim=c(2,5))+
  facet_grid(~Q409) +
  scale_x_discrete(labels = c("Pre", "Post")) +
  scale_color_manual(values=c("#0072B2", "#D55E00" )) +
  scale_fill_manual(values=c("#0072B2","#D55E00"))

ggsave(LYliking,file='../figures/LYliking.png', width=9, dpi=1000)
```

#analyses
```{r}
#did the s2s illusion work in entire sample?
m2 <- lmer(data=speech_master, scale(value) ~ stim_type*pre_post +  (stim_type*pre_post|ResponseId) + (pre_post|stim), control=lmerControl(optimizer="bobyqa"))

summary(m1)
test(emmeans(m1, ~ stim_type*pre_post, var = "stim_type", lmer.df = "satterthwaite",lmerTest.limit = 85652))


m2 <- lmer(data=liking_master, scale(value) ~ stim_type*pre_post +  (stim_type*pre_post|ResponseId) + (pre_post|stim), control=lmerControl(optimizer="bobyqa"))
summary(m2)

#how about for anhedonics?
#speech
m1a <- lmer(data=speech_matched, scale(value) ~ stim_type*group + (group|ResponseId),control=lmerControl(optimizer="bobyqa"))
summary(m1a)
#emmeans(m1a, ~stim_type*group,lmerTest.limit = 4224)

m1b <- lmer(data=speech_matched, scale(value) ~ group*pre_post + (group|ResponseId),control=lmerControl(optimizer="bobyqa"))
summary(m1b)

#liking
m2a <- lmer(data=liking_matched, scale(value) ~ stim_type*group + (group|ResponseId),control=lmerControl(optimizer="bobyqa"))
summary(m2a)

m2b <- lmer(data=speech_matched, scale(value) ~ group*pre_post + (group|ResponseId) + (1|stim),control=lmerControl(optimizer="bobyqa"))
summary(m2b)

#and laurel v yanny hearers
m1c <- lmer(data=speech_master, scale(value) ~ Q409*pre_post*stim_type + (Q409|ResponseId), control=lmerControl(optimizer="bobyqa"))
summary(m1c)

m2c <- lmer(data=liking_master, scale(value) ~ Q409*pre_post*stim_type + (1|ResponseId), control=lmerControl(optimizer="bobyqa"))
summary(m2c)

#plot <-
speech_master %>%
  subset(location=="boston") %>%
  ggplot(data=, aes(x=pre_post, y=value, color=stim_type)) + 
 #ggdist::geom_dotsinterval() +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'errorbar', width = 0, alpha=0.5) +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'pointrange', alpha=0.5) +
  theme_classic()+ 
  #geom_jitter(width=0.4,height=0.4,alpha=0.1) +
  #geom_line(alpha=0.05, color="black")+
  #geom_point()+
 #theme(legend.position = "none")+
  ylab("Musicality Rating") +
  xlab("Pre/Post Repetition") +
  theme(
    axis.text.x = element_text(family="Arial", size=12),
    axis.title.x = element_text(family="Arial", size=12),
    axis.title.y = element_text(family="Arial", size=12)) +
    scale_x_discrete(labels = c("Pre Repetition", "Post Reptition")) +
    coord_cartesian(ylim=c(2,5))+
  facet_grid(~Q409) 

ggsave(plot,file='../figures/prelim/LY_speechprepost.png')


m12 <- lmer(data=liking_master, scale(value) ~ Q409*pre_post + (Q409|ResponseId), control=lmerControl(optimizer="bobyqa"))
summary(m12)

m12 <- lmer(data=liking_master, scale(value) ~ Q409*stim_type + (Q409|ResponseId), control=lmerControl(optimizer="bobyqa"))
summary(m12)
```


```{r}
#are liking + musicality ratings linked? 
## maybe test if this linkage relates to eBMRQ?

#are pre/post changes related?
liking_master_merge <- liking_master_post %>%
  select(ResponseId, stim, liking_change, stim_type)

speech_master_merge <- speech_master_post %>%
  select(ResponseId, stim, musicality_change, stim_type)


change_master <- merge(liking_master_merge, speech_master_merge, by=c("ResponseId","stim", "stim_type"))

cor.test(change_master$liking_change, change_master$musicality_change)

m9 <- lmer(data=change_master, scale(liking_change) ~ scale(musicality_change)*stim_type + (scale(musicality_change)*stim_type|ResponseId) + (1|stim),control=lmerControl(optimizer="bobyqa"))
summary(m9)

#adding model predictions
newdata <- expand.grid(
  musicality = seq(-9,9,0.01),
  pre_post = c("pre", "post"),
  liking= seq(4.487640,6.284753, 0.01)
)

newdata$pre_post <- as.factor(newdata$pre_post)
newdata$pre_post <- factor(newdata$pre_post, levels = c("pre", "post"))
contrasts(newdata$pre_post) <- c(-0.5,0.5)

predictions<- AICcmodavg::predictSE(m9,newdata, se.fit = TRUE, print.matrix =
        TRUE, level = 0)

pred2a <- cbind(newdata, predictions)

ggplot(data = change_master, aes(x=musicality_change, y=liking_change)) +
  geom_jitter(width=0.1, height=0.1, alpha=0.2) +
  theme_classic() +
  stat_smooth(method="lm") +
  xlab("Change in Musicality Rating Post-Repetition") +
  ylab("Change in Liking Rating Post-Repetition")

master_long$musicality_scaled <- scale(master_long$musicality)
master_long$liking_scaled <- scale(master_long$liking)

m9 <- lmer(data=master_long, liking ~ musicality*pre_post + (musicality*pre_post|ResponseId) + (1|stim),control=lmerControl(optimizer="bobyqa"))
summary(m9)

test(emmeans::emtrends(m9, pairwise ~ pre_post, var = "musicality_scaled", lmer.df = "satterthwaite",lmerTest.limit = 85652))

#plotting below, first making a df w/ means + SEs for liking + musicality per stimulus
#getting liking SEs to plot below:
#musicality
musicality_means_pre <- 
speech_master_pre %>%
  #subset(pre_post=="pre") %>%
  group_by(stim, pre_post) %>%
  dplyr::summarize(musicality = mean(value, na.rm=TRUE)) %>%
  drop_na()

SE_musicality_pre <-
  speech_master_pre %>%
  subset(pre_post=="pre") %>%
  group_by(stim, pre_post) %>%
  dplyr::summarize(musicality_sd = sd(value, na.rm=TRUE)) %>%
  drop_na()


SE_musicality_pre$freq <-table(speech_master_pre$stim)

SE_musicality_pre$SE_musicality <- (SE_musicality_pre$musicality_sd)/(sqrt(SE_musicality_pre$freq))

musicality_means_post <- 
speech_master_post %>%
  subset(pre_post=="post") %>%
  group_by(stim, pre_post) %>%
  dplyr::summarize(musicality = mean(value, na.rm=TRUE)) %>%
  drop_na() 

SE_musicality_post <-
  speech_master_post %>%
  subset(pre_post=="post") %>%
  group_by(stim, pre_post) %>%
  dplyr::summarize(musicality_sd = sd(value, na.rm=TRUE)) %>%
  drop_na() 

SE_musicality_post$freq <-table(speech_master_post$stim)

SE_musicality_post$SE_musicality <- (SE_musicality_post$musicality_sd)/(sqrt(SE_musicality_post$freq))

SE_musicality_post <- merge(SE_musicality_post, musicality_means_post, by=c("stim", "pre_post"))

SE_musicality_pre <- merge(SE_musicality_pre, musicality_means_pre, by=c("stim", "pre_post"))

SE_musicality <- rbind(SE_musicality_post, SE_musicality_pre)

SE_musicality$lower <- SE_musicality$musicality-SE_musicality$SE_musicality
SE_musicality$upper <- SE_musicality$musicality+SE_musicality$SE_musicality

#liking
liking_means_pre <- 
liking_master_pre %>%
  subset(pre_post=="pre") %>%
  group_by(stim, pre_post) %>%
  dplyr::summarize(liking = mean(value, na.rm=TRUE)) %>%
  drop_na() 

SE_liking_pre <-
  liking_master_pre %>%
  subset(pre_post=="pre") %>%
  group_by(stim, pre_post) %>%
  dplyr::summarize(liking_sd = sd(value, na.rm=TRUE)) %>%
  drop_na()

SE_liking_pre$freq <- table(liking_master_pre$stim)

SE_liking_pre$SE_liking <- (SE_liking_pre$liking_sd)/(sqrt(SE_liking_pre$freq))

liking_means_post <- 
liking_master_post %>%
  subset(pre_post=="post") %>%
  group_by(stim, pre_post) %>%
  dplyr::summarize(liking = mean(value, na.rm=TRUE)) %>%
  drop_na()

SE_liking_post <-
  liking_master_post %>%
  subset(pre_post=="post") %>%
  group_by(stim, pre_post) %>%
  dplyr::summarize(liking_sd = sd(value, na.rm=TRUE)) %>%
  drop_na() 

SE_liking_post$freq <- table(liking_master_post$stim)

SE_liking_post$SE_liking <- (SE_liking_post$liking_sd)/(sqrt(SE_liking_post$freq))

SE_liking_post <- merge(SE_liking_post, liking_means_post, by=c("stim", "pre_post")) 


SE_liking_pre <- merge(SE_liking_pre, liking_means_pre, by=c("stim","pre_post")) 

SE_liking <- rbind(SE_liking_post, SE_liking_pre)

SE_liking$lower_liking <- SE_liking$liking-SE_liking$SE_liking
SE_liking$upper_liking <- SE_liking$liking+SE_liking$SE_liking

SE <- merge(SE_liking, SE_musicality, by=c("stim", "pre_post"))
SE$pre_post <- factor(SE$pre_post, levels = c("pre", "post"))

#adding model predictions
newdata <- expand.grid(
  musicality = seq(1.810962,5.653244,0.01),
  pre_post = c("pre", "post"),
  liking= seq(4.487640,6.284753, 0.01)
)

newdata$pre_post <- as.factor(newdata$pre_post)
newdata$pre_post <- factor(newdata$pre_post, levels = c("pre", "post"))
contrasts(newdata$pre_post) <- c(-0.5,0.5)

predictions<- AICcmodavg::predictSE(m9,newdata, se.fit = TRUE, print.matrix =
        TRUE, level = 0)

pred2a <- cbind(newdata, predictions)


#unscaling
#pred2a$musicality <- pred2a$musicality_scaled * attr(master_long$musicality_scaled, 'scaled:scale') + attr(master_long$musicality_scaled, 'scaled:center')

#unscaling predictions + adding upper/lower SE
pred2a$liking <- pred2a$fit #* attr(master_long$liking_scaled, 'scaled:scale') + attr(master_long$liking_scaled, 'scaled:center')
pred2a$upper <- (pred2a$fit + pred2a$se.fit) #* attr(master_long$liking_scaled, 'scaled:scale') + attr(master_long$liking_scaled, 'scaled:center')
pred2a$lower <- (pred2a$fit - pred2a$se.fit) #* attr(master_long$liking_scaled, 'scaled:scale') + attr(master_long$liking_scaled, 'scaled:center')

SE$labels <- factor(SE$pre_post, levels = c("pre", "post"),
                  labels = c("Pre Repetition", "Post Repetition")
                  )
#something weird going on in plot here
ggplot(data=SE, aes(x=musicality, y=liking))+
  geom_point(alpha=0.5)+
  #stat_summary(fun.data  = 'mean_se', geom = 'errorbar', width = 0, alpha=0.5) +
  #stat_summary(fun.data = 'mean_se', geom = 'pointrange', alpha=0.5)+
  theme_classic()+
 geom_errorbar(data=SE, aes(ymin = lower_liking, ymax = upper_liking), alpha=0.2, width=0)+
geom_errorbarh(data=SE, aes(xmin = lower, xmax = upper), alpha=0.2,  height=0) +
 geom_line(data=pred2a, color="black") +
 geom_ribbon(data=pred2a, aes(ymin=upper, ymax=lower),alpha=0.2)  +
  facet_grid(~labels) +
  xlab("Musicality Rating") +
  ylab("Liking Rating")
  
```

```{r}


#just illusion stims, prev post
mode_liking <- 
  liking_matched %>%
  subset(stim_type=="illusion")

model <- lmer(data = mode_liking, scale(value) ~ pre_post*group + (1|ResponseId) + (1|stim),control=lmerControl(optimizer="bobyqa"))

summary(model)

#plot <- 
liking_matched$labels <- factor(liking_matched$pre_post, levels = c("pre", "post"),
                  labels = c("Pre Repetition", "Post Repetition")
                  )
liking_matched$stim_type <- ifelse(liking_matched$stim_type=="control", "Control", "Illusion")


ggplot(data=liking_matched, aes(x=stim_type, y=value, color=group))+ 
 # ggdist::geom_dotsinterval() +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'errorbar', width = 0, alpha=0.5) +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'pointrange', alpha=0.5) +
  theme_classic()+ 
  #geom_jitter(width=0.2,height=0.2,alpha=0.2) +
  #geom_line(alpha=0.05, color="black")+
  theme(legend.position = "none")+
  xlab("Pre/Post Repetition") +
  ylab("Liking Rating") +
 scale_color_manual(values=c("#56B4E9", "#E69F00" )) +
  scale_fill_manual(values=c("#56B4E9","#E69F00")) +
  facet_grid(~labels)

ggsave(plot,file='../figures/prelim/likingraw_MAMC.png')

#faceted plot
#liking_musciality_all <-
 
  #

#plotting three-way interaction
liking_matched %>%
  subset(location=="birkbeck") %>%
  ggplot(aes(x=labels, y=value, color=group)) + 
 # ggdist::geom_dotsinterval() +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'errorbar', width = 0, alpha=0.5) +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'pointrange', alpha=0.5) +
  theme_bw()+
  theme(legend.position = "none")+
  xlab("Stimulus Type") +
  ylab("Liking Rating") +
 scale_color_manual(values=c("#56B4E9", "#E69F00" )) +
 scale_fill_manual(values=c("#56B4E9","#E69F00")) +
  facet_grid(~stim_type)+
  coord_cartesian(ylim=c(3,7))


#plotting raw
ggplot(data=liking_matched, aes(x=pre_post, y=value, color=stim_type)) + 
 #ggdist::geom_dotsinterval() +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'errorbar', width = 0, alpha=0.5) +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'pointrange', alpha=0.5) +
  theme_classic()+ 
  #geom_jitter(width=0.4,height=0.4,alpha=0.1) +
  #geom_line(alpha=0.05, color="black")+
  #geom_point()+
 theme(legend.position = "none")+
  ylab("Liking Rating") +
  xlab("Pre vs. Post Repetition") +
  facet_grid(~group)
  #scale_color_manual(values=c("#56B4E9", "#E69F00" )) +
  #scale_fill_manual(values=c("#56B4E9","#E69F00"))

cor.test(liking_matchedprepost$change, speech_matchedprepost$change)

#plotting average post musicality ratings per group per stimulus
plot<-
  ggplot(data=speech_matchedprepost, aes(x=stim, y=post, color=group)) + 
 #ggdist::geom_dotsinterval() +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'errorbar', width = 0, alpha=0.5) +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'pointrange', alpha=0.5) +
  theme_classic()+ 
  #geom_jitter(width=0.4,height=0.4,alpha=0.1) +
  #geom_line(alpha=0.05, color="black")+
  #geom_point()+
 theme(legend.position = "none")+
  ylab("Post-Repetition Musicality Rating") +
  xlab("Stimulus") +
  scale_color_manual(values=c("#56B4E9", "#E69F00" )) +
  scale_fill_manual(values=c("#56B4E9","#E69F00")) +
 scale_x_continuous(breaks = speech_matchedprepost$stim) +
  theme(
    axis.text.x = element_text(family="Arial", size=12),
    axis.title.x = element_text(family="Arial", size=12),
    axis.title.y = element_text(family="Arial", size=12)) +
ggsave(plot,file='../figures/prelim/musicalityperstim_MAMC.png', width=14)

#plot<-
  ggplot(data=liking_matchedprepost, aes(x=stim, y=post, color=group)) + 
 #ggdist::geom_dotsinterval() +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'errorbar', width = 0, alpha=0.5) +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'pointrange', alpha=0.5) +
  theme_classic()+ 
  #geom_jitter(width=0.4,height=0.4,alpha=0.1) +
  #geom_line(alpha=0.05, color="black")+
  #geom_point()+
 theme(legend.position = "none")+
  ylab("Post-Repetition Liking Rating") +
  xlab("Stimulus") +
  scale_color_manual(values=c("#56B4E9", "#E69F00" )) +
  scale_fill_manual(values=c("#56B4E9","#E69F00")) +
 scale_x_continuous(breaks = speech_matchedprepost$stim) +
  theme(
    axis.text.x = element_text(family="Arial", size=12),
    axis.title.x = element_text(family="Arial", size=12),
    axis.title.y = element_text(family="Arial", size=12))
ggsave(plot,file='../figures/prelim/perstim_MAMC.png', width=14)

#yanny v laurel hearers

liking_master %>%
  subset()
ggplot(data=, aes(x=pre_post, y=value, color=Q409)) + 
 #ggdist::geom_dotsinterval() +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'errorbar', width = 0, alpha=0.5) +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'pointrange', alpha=0.5) +
  theme_classic()+ 
  #geom_jitter(width=0.4,height=0.4,alpha=0.1) +
  #geom_line(alpha=0.05, color="black")+
  #geom_point()+
 theme(legend.position = "none")+
  ylab("Liking Rating") +
  xlab("Pre/Post Repetition") +
  theme(
    axis.text.x = element_text(family="Arial", size=12),
    axis.title.x = element_text(family="Arial", size=12),
    axis.title.y = element_text(family="Arial", size=12))+
    scale_x_discrete(labels = c("Pre Repetition", "Post Reptition")) +
  facet_grid(~stim_type)


check <- 
  speech_master %>%
  distinct(Q409,ResponseId)
table(check$Q409)
ggplot(data=liking_master, aes(x=pre_post, y=value, color=Q409)) + 
 #ggdist::geom_dotsinterval() +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'errorbar', width = 0, alpha=0.5) +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'pointrange', alpha=0.5) +
  theme_classic()+ 
  #geom_jitter(width=0.4,height=0.4,alpha=0.1) +
  #geom_line(alpha=0.05, color="black")+
  #geom_point()+
 #theme(legend.position = "none")+
  ylab("Liking Rating") +
  xlab("Pre/Post Repetition") +
  theme(
    axis.text.x = element_text(family="Arial", size=12),
    axis.title.x = element_text(family="Arial", size=12),
    axis.title.y = element_text(family="Arial", size=12))

#checking illusion v control, laurel v yanny - post
speech_master_post <- speech_master %>%
  subset(pre_post == "post")

m12 <- lmer(data=speech_master_post, scale(value) ~ Q409*stim_type + (Q409+stim_type|ResponseId), control=lmerControl(optimizer="bobyqa"))
summary(m12)

m13 <- lmer(data=speech_master, scale(value) ~ Q409*pre_post + (Q409*pre_post|ResponseId), control=lmerControl(optimizer="bobyqa"))
summary(m13)

#plot <- 
ggplot(data=speech_master_post, aes(x=stim_type, y=value, color=Q409)) + 
 #ggdist::geom_dotsinterval() +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'errorbar', width = 0, alpha=0.5) +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'pointrange', alpha=0.5) +
  theme_classic()+ 
  #geom_jitter(width=0.4,height=0.4,alpha=0.1) +
  #geom_line(alpha=0.05, color="black")+
  #geom_point()+
 #theme(legend.position = "none")+
  ylab("Post Musicality Rating") +
  xlab("Stimulus Type") +
  theme(
    axis.text.x = element_text(family="Arial", size=12),
    axis.title.x = element_text(family="Arial", size=12),
    axis.title.y = element_text(family="Arial", size=12))

ggsave(plot,file='../figures/prelim/LY_musicalitypoststim.png')



ggplot(data=speech_master, aes(x=pre_post, y=value, color=Q409)) + 
 #ggdist::geom_dotsinterval() +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'errorbar', width = 0, alpha=0.5) +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'pointrange', alpha=0.5) +
  theme_classic()+ 
  #geom_jitter(width=0.4,height=0.4,alpha=0.1) +
  #geom_line(alpha=0.05, color="black")+
  #geom_point()+
  theme(legend.position = "none")+
  ylab("Musicality Rating") +
  xlab("") +
  #theme(
  #  axis.text.x = element_text(family="Arial", size=12),
   # axis.title.x = element_text(family="Arial", size=12),
    #axis.title.y = element_text(family="Arial", size=12)) +
  facet_grid(~stim_type) +
    scale_x_discrete(labels = c("Pre Repetition", "Post Reptition"))


ggplot(data=liking_master, aes(x=pre_post, y=value, color=Q409)) + 
 #ggdist::geom_dotsinterval() +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'errorbar', width = 0, alpha=0.5) +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'pointrange', alpha=0.5) +
  theme_classic()+ 
  #geom_jitter(width=0.4,height=0.4,alpha=0.1) +
  #geom_line(alpha=0.05, color="black")+
  #geom_point()+
  theme(legend.position = "none")+
  ylab("Liking Rating") +
  xlab("") +
  #theme(
  #  axis.text.x = element_text(family="Arial", size=12),
   # axis.title.x = element_text(family="Arial", size=12),
    #axis.title.y = element_text(family="Arial", size=12)) +
  facet_grid(~stim_type) +
    scale_x_discrete(labels = c("Pre Repetition", "Post Reptition"))


musicality_averages <-
  speech_master %>%
  group_by(pre_post,stim_type, ResponseId) %>%
  dplyr::summarize(musicality = mean(musicality, na.rm=TRUE)) %>%
  drop_na()

musicality_sd <-
  speech_master %>%
  group_by(pre_post,stim_type, ResponseId) %>%
  dplyr::summarize(musicality_sd = sd(musicality, na.rm=TRUE)) %>%
  drop_na()


musicality_sd$SE_liking <- (musicality_sd$musicality_sd)/(sqrt(24))

SE_liking_post <- merge(SE_liking_post, liking_means_post, by=c("stim", "pre_post"))


MAMC_musicality <-
speech_matched %>%
  mutate(rating = recode(rating, "liking" = "Liking", "speech" = "Musicality")) %>%
  #mutate(rating = factor(rating, levels = c("Musicality", "Liking"))) %>%
  mutate(pre_post = factor(pre_post, levels = c("pre", "post"))) %>%
  mutate(stim_type = recode(stim_type, "illusion" = "Illusion", "control" = "Control")) %>%
  mutate(group = factor(group, levels = c("Musical Anhedonics", "Matched Controls"))) %>%
  ggplot(aes(x=pre_post, y=value, color=stim_type)) + 
 #ggdist::geom_dotsinterval() +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'errorbar', width = 0.05, alpha=0.5) +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'pointrange', alpha=0.5) +
  theme_bw()+ 
  #geom_jitter(width=0.4,height=0.4,alpha=0.1) +
  #geom_line(alpha=0.05, color="black")+
  #geom_point()+
 theme(legend.position = "none")+
  ylab("Musicality Rating") +
  xlab("") +
  scale_x_discrete(labels = c("Pre", "Post")) +
  #coord_cartesian(ylim = c(2,5)) +
   theme(
    axis.title.x = element_text(family="Arial", size=20,face="bold"),
    axis.title.y = element_text(family="Arial",size=20,face="bold"),
    axis.text.x = element_text(family="Arial", size=14,face="bold"),
    axis.text.y = element_text(family="Arial",size=14,face="bold"),
    strip.text = element_text(family="Arial",size=16,color='black',face="bold")
  ) +
  facet_wrap(~group) 
 
ggsave(MAMC_musicality,file='../figures/prelim/MAMC_musicality.png', width=9)


MAMC_liking <-
liking_matched %>%
  mutate(rating = recode(rating, "liking" = "Liking", "speech" = "Musicality")) %>%
  #mutate(rating = factor(rating, levels = c("Musicality", "Liking"))) %>%
  mutate(pre_post = factor(pre_post, levels = c("pre", "post"))) %>%
  mutate(stim_type = recode(stim_type, "illusion" = "Illusion", "control" = "Control")) %>%
  mutate(group = factor(group, levels = c("Musical Anhedonics", "Matched Controls"))) %>%
  ggplot(aes(x=pre_post, y=value, color=stim_type)) + 
 #ggdist::geom_dotsinterval() +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'errorbar', width = 0.05, alpha=0.5) +
  stat_summary(fun.data = 'mean_cl_normal', geom = 'pointrange', alpha=0.5) +
  theme_bw()+ 
  #geom_jitter(width=0.4,height=0.4,alpha=0.1) +
  #geom_line(alpha=0.05, color="black")+
  #geom_point()+
 theme(legend.position = "none")+
  ylab("Liking Rating") +
  xlab("") +
  scale_x_discrete(labels = c("Pre", "Post")) +
  #coord_cartesian(ylim = c(2,5)) +
   theme(
    axis.title.x = element_text(family="Arial", size=20,face="bold"),
    axis.title.y = element_text(family="Arial",size=20,face="bold"),
    axis.text.x = element_text(family="Arial", size=14,face="bold"),
    axis.text.y = element_text(family="Arial",size=14,face="bold"),
    strip.text = element_text(family="Arial",size=16,color='black',face="bold")
  ) +
  facet_wrap(~group) 
 
ggsave(MAMC_liking,file='../figures/prelim/MAMC_liking.png', width=9)

```

#testing beta coefs
```{r}
#musicality
betas_musicality <- coef(m1)$ResponseId %>% 
    rownames_to_column()

colnames(betas_musicality) <- c("ResponseId", "Intercept_musicality", "stim_type_musicality", "pre_post_musicality", "pre_postXstim_type_musicality")

participant_BMRQs <- unique(dplyr::select(master, c("ResponseId", "eBMRQ_total", "BMRQ_sensorimotor", "BMRQ_moodreg", "BMRQ_emotion", "BMRQ_musicseek", "BMRQ_socialreward", "BMRQ_absorption", "AIMS_total")))


matched_groups <- unique(dplyr::select(speech_matched, c("ResponseId", "group")))

betas_merged <- merge(betas_musicality, participant_BMRQs, on="ResponseId")

betas_merged_matched <- merge(betas_musicality, matched_groups, on="ResponseId") 

anhedonics_beta <- subset(betas_merged_matched, group=="Musical Anhedonics")
controls_beta <- subset(betas_merged_matched, group=="Matched Controls")

model1 <- lm(data=betas_merged, scale(pre_post_musicality) ~ scale(eBMRQ_total))
summary(model1)

model2 <- lm(data=betas_merged, scale(stim_type_musicality) ~ scale(eBMRQ_total))
summary(model2)

model3 <- lm(data=betas_merged, scale(pre_postXstim_type_musicality) ~ scale(eBMRQ_total))
summary(model3)

t.test(anhedonics_beta$pre_postXstim_type_musicality, controls_beta$pre_postXstim_type_musicality)

#liking
betas_liking <- coef(m2)$ResponseId %>% 
    rownames_to_column()

colnames(betas_liking) <- c("ResponseId", "Intercept_liking", "stim_type_liking", "pre_post_liking", "pre_postXstim_type_liking")

betas_merged <- merge(betas_liking, betas_merged, on="ID")

betas_merged_matched <- merge(betas_liking, betas_merged_matched, on="ID")


model1 <- lm(data=betas_merged, scale(pre_post_liking) ~ scale(eBMRQ_total))
summary(model1)

model2 <- lm(data=betas_merged, scale(stim_type_liking) ~ scale(eBMRQ_total))
summary(model2)

model3 <- lm(data=betas_merged, scale(pre_postXstim_type_liking) ~ scale(eBMRQ_total))
summary(model3)

t.test(anhedonics_beta$pre_postXstim_type_liking, controls_beta$pre_postXstim_type_liking)

#are liking + musicality effects related?
model1 <- lm(data=betas_merged, scale(pre_post_liking) ~ scale(pre_post_musicality))
summary(model1)

model2 <- lm(data=betas_merged, scale(stim_type_liking) ~ scale(stim_type_musicality))
summary(model2)

model3 <- lm(data=betas_merged, scale(pre_postXstim_type_liking) ~ scale(pre_postXstim_type_musicality))
summary(model3)

#plotting
#eBMRQ + effect of repetition --> how to deal w/ outliers?
ggplot(data=betas_merged, aes(x=pre_post_musicality, y=eBMRQ_total)) +
  geom_point()+
  stat_smooth(method="lm") +
  ylab("Total eBMRQ Score") +
  xlab("Estimated Effect of Repetition on Musicality Ratings") +
  theme_classic()

ggplot(data=betas_merged, aes(x=stim_type_musicality, y=eBMRQ_total)) +
  geom_point()+
  stat_smooth(method="lm") +
  ylab("Total eBMRQ Score") +
  xlab("Estimated Effect of Stimulus Type on Musicality Ratings") +
  theme_classic()

ggplot(data=betas_merged, aes(x=pre_postXstim_type_musicality, y=eBMRQ_total)) +
  geom_point()+
  stat_smooth(method="lm") +
  ylab("Total eBMRQ Score") +
  xlab("Estimated Repetition X Stimulus Type Interaction on Musicality Ratings") +
  theme_classic()

#still need to do liking ratings + both w/ AIMS scores, though currently all are n.s.

#relationship between effects
ggplot(data=betas_merged, aes(x=pre_post_musicality, y=pre_post_liking)) +
  geom_point()+
  stat_smooth(method="lm") +
  ylab("Estimated Effect of Repetition on\nLiking Ratings") +
  xlab("Estimated Effect of Repetition on Musicality Ratings") +
  theme_classic()

betas_merged_no_outlier<-betas_merged %>%
  filter(ResponseId != "R_8EGYGuIqac9niiq")

ggplot(data=betas_merged , aes(x=stim_type_musicality, y=stim_type_liking)) +
  geom_point()+
  stat_smooth(method="lm") +
  ylab("Estimated Effect of Stimulus Type on\nLiking Ratings") +
  xlab("Estimated Effect of Stimulus Type on Musicality Ratings") +
  theme_classic()

ggplot(data=betas_merged_no_outlier , aes(x=stim_type_musicality, y=stim_type_liking)) +
  geom_point()+
  stat_smooth(method="lm") +
  ylab("Estimated Effect of Stimulus Type on\nLiking Ratings") +
  xlab("Estimated Effect of Stimulus Type on Musicality Ratings") +
  theme_classic()

ggplot(data=betas_merged, aes(x=pre_postXstim_type_musicality, y=pre_postXstim_type_liking)) +
  geom_point()+
  stat_smooth(method="lm") +
  ylab("Estimated Repetition X Stimulus Type\nInteraction on Liking Ratings") +
  xlab("Estimated Repetition X Stimulus Type Interaction on Musicality Ratings") +
  theme_classic()

ggplot(data=betas_merged_no_outlier, aes(x=pre_postXstim_type_musicality, y=pre_postXstim_type_liking)) +
  geom_point()+
  stat_smooth(method="lm") +
  ylab("Estimated Repetition X Stimulus Type\nInteraction on Liking Ratings") +
  xlab("Estimated Repetition X Stimulus Type Interaction on Musicality Ratings") +
  theme_classic()
```

#individual differences
```{r}
#did musicality ratings differ by eBMRQ scores?
#speech_masterprepost$stim_type <- as.factor(speech_masterprepost$stim_type)
#contrasts(speech_masterprepost$stim_type) <- c(-0.5,0.5)
m3 <- lmer(data=speech_master, scale(value) ~ stim_type*scale(eBMRQ_total)*pre_post + (stim_type*scale(eBMRQ_total) + scale(eBMRQ_total)*pre_post + stim_type*pre_post|ResponseId) + (1|stim),control=lmerControl(optimizer="bobyqa"))
summary(m3)

test(emmeans::emtrends(m3, pairwise ~ pre_post*stim_type, var = "eBMRQ_total", lmer.df = "satterthwaite",lmerTest.limit = 85652))

#adding model predictions
newdata <- expand.grid(
  eBMRQ_total = seq(27,120,1),
  pre_post = c("pre", "post"),
  stim_type= c("control", "illusion"))

newdata$pre_post <- as.factor(newdata$pre_post)
newdata$pre_post <- factor(newdata$pre_post, levels = c("pre", "post"))
contrasts(newdata$pre_post) <- c(-0.5,0.5)
contrasts(newdata$stim_type) <- c(-0.5,0.5)

predictions<- AICcmodavg::predictSE(m3,newdata, se.fit = TRUE, print.matrix =
        TRUE, level = 0)

pred2a <- cbind(newdata, predictions)

speech_master$scaled_musicality <- scale(speech_master$value)

#unscaling
pred2a$musicality <- pred2a$fit * attr(speech_master$scaled_musicality, 'scaled:scale') + attr(speech_master$scaled_musicality, 'scaled:center')

#unscaling predictions + adding upper/lower SE
pred2a$upper <- (pred2a$fit + pred2a$se.fit) * attr(speech_master$scaled_musicality, 'scaled:scale') + attr(speech_master$scaled_musicality, 'scaled:center')
pred2a$lower <- (pred2a$fit - pred2a$se.fit) * attr(speech_master$scaled_musicality, 'scaled:scale') + attr(speech_master$scaled_musicality, 'scaled:center')

ggplot(data=speech_master , aes(x=eBMRQ_total, y=value)) +
  #stat_summary(fun.data = 'mean_se', geom = 'errorbar', width = 0, alpha=0.5) +
  #stat_summary(fun.data = 'mean_se', geom = 'pointrange', alpha=0.5)+
  #stat_smooth(method="lm") +
  ylab("Musicality Rating") +
  xlab("Total eBMRQ Score") +
  theme_classic() +
  facet_grid(pre_post~stim_type) +
  geom_line(data=pred2a, aes(y=musicality, x=eBMRQ_total, color=pre_post)) +
  geom_ribbon(data=pred2a, aes(y=musicality, ymin=upper, ymax=lower,color=pre_post, fill=pre_post),alpha=0.2) +
  facet_grid(~stim_type)

m33 <- lmer(data=speech_master, scale(value) ~ stim_type*scale(AIMS_total)*pre_post + (stim_type*scale(AIMS_total) + scale(AIMS_total)*pre_post + stim_type*pre_post|ResponseId) + (1|stim),control=lmerControl(optimizer="bobyqa"))
summary(m33)

test(emmeans::emtrends(m33,pairwise ~ pre_post*stim_type, var = "AIMS_total", lmer.df = "satterthwaite",lmerTest.limit = 85652))

#adding model predictions
newdata <- expand.grid(
  AIMS_total = seq(35,170,1),
  pre_post = c("pre", "post"),
  stim_type= c("control", "illusion"))

newdata$pre_post <- as.factor(newdata$pre_post)
newdata$pre_post <- factor(newdata$pre_post, levels = c("pre", "post"))
contrasts(newdata$pre_post) <- c(-0.5,0.5)
contrasts(newdata$stim_type) <- c(-0.5,0.5)

predictions<- AICcmodavg::predictSE(m33,newdata, se.fit = TRUE, print.matrix =
        TRUE, level = 0)

pred2a <- cbind(newdata, predictions)

#unscaling
pred2a$musicality <- pred2a$fit * attr(speech_master$scaled_musicality, 'scaled:scale') + attr(speech_master$scaled_musicality, 'scaled:center')

#unscaling predictions + adding upper/lower SE
pred2a$upper <- (pred2a$fit + pred2a$se.fit) * attr(speech_master$scaled_musicality, 'scaled:scale') + attr(speech_master$scaled_musicality, 'scaled:center')
pred2a$lower <- (pred2a$fit - pred2a$se.fit) * attr(speech_master$scaled_musicality, 'scaled:scale') + attr(speech_master$scaled_musicality, 'scaled:center')

ggplot(data=speech_master , aes(x=AIMS_total, y=value, color=pre_post)) +
  #stat_summary(fun.data = 'mean_se', geom = 'errorbar', width = 0, alpha=0.5) +
  #stat_summary(fun.data = 'mean_se', geom = 'pointrange', alpha=0.5)+
  #stat_smooth(method="lm") +
  #geom_jitter(height=0.2, alpha=0.2) +
  ylab("Musicality Rating") +
  xlab("Total AIMS Score") +
  theme_classic() +
  facet_grid(~stim_type) +
  geom_line(data=pred2a, aes(y=musicality, x=AIMS_total,color=pre_post)) +
  geom_ribbon(data=pred2a, aes(y=musicality, ymin=upper, ymax=lower,color=pre_post, fill=pre_post),alpha=0.2)



m4 <- lmer(data=speech_master, scale(value) ~ pre_post*scale(eBMRQ_total)*stim_type + (pre_post*scale(eBMRQ_total)*stim_type|ResponseId) + (1|stim),control=lmerControl(optimizer="bobyqa"))
summary(m4)

m5 <- lmer(data=speech_master, value ~ pre_post*scale(eBMRQ_total) + (pre_post*scale(eBMRQ_total)|ResponseId) + (1|stim),control=lmerControl(optimizer="bobyqa"))
summary(m5)

#adding model predictions
newdata <- expand.grid(
  eBMRQ_total = seq(27,120,1),
  pre_post = c("pre", "post"))

newdata$pre_post <- as.factor(newdata$pre_post)
newdata$pre_post <- factor(newdata$pre_post, levels = c("pre", "post"))
contrasts(newdata$pre_post) <- c(-0.5,0.5)

predictions<- AICcmodavg::predictSE(m5,newdata, se.fit = TRUE, print.matrix =
        TRUE, level = 0)

pred2a <- cbind(newdata, predictions)


#unscaling
pred2a$musicality <- pred2a$musicality_scaled * attr(master_long$musicality_scaled, 'scaled:scale') + attr(master_long$musicality_scaled, 'scaled:center')

#unscaling predictions + adding upper/lower SE
pred2a$value <- pred2a$fit #* attr(master_long$liking_scaled, 'scaled:scale') + attr(master_long$liking_scaled, 'scaled:center')
pred2a$upper <- (pred2a$fit + pred2a$se.fit) #* attr(master_long$liking_scaled, 'scaled:scale') + attr(master_long$liking_scaled, 'scaled:center')
pred2a$lower <- (pred2a$fit - pred2a$se.fit) #* attr(master_long$liking_scaled, 'scaled:scale') + attr(master_long$liking_scaled, 'scaled:center')

#change pre/post to pre-repeition and post-repetition for facet titles
ggplot(data=speech_master, aes(x=eBMRQ_total, y=value))+
  #geom_point()+
  #stat_summary(fun.data = 'mean_se', geom = 'errorbar', width = 0, alpha=0.5) +
  #stat_summary(fun.data = 'mean_se', geom = 'pointrange', alpha=0.5)+
  theme_classic()+
  #geom_smooth(method="lm") +
  #facet_grid(~stim_type) +
  xlab("Total eBMRQ Score") +
  ylab("Musicality Rating") +
  geom_line(data=pred2a) +
  geom_ribbon(data=pred2a, aes(ymin=upper, ymax=lower),alpha=0.2) +
  facet_grid(~pre_post)
  


#did liking ratings differ by eBMRQ scores?
liking_master$stim_type <- as.factor(liking_master$stim_type)
contrasts(liking_master$stim_type) <- c(-0.5,0.5)

m4 <- lmer(data=liking_master, scale(value) ~ stim_type*scale(eBMRQ_total)*pre_post + (stim_type:scale(eBMRQ_total):pre_post + stim_type:scale(eBMRQ_total) + pre_post:stim_type + pre_post*scale(eBMRQ_total)|ResponseId) + (1|stim),control=lmerControl(optimizer="bobyqa"))

summary(m4)

m44 <- lmer(data=liking_master, scale(value) ~ stim_type*scale(AIMS_total)*pre_post + (stim_type:scale(AIMS_total):pre_post + stim_type:scale(AIMS_total) + pre_post:stim_type + pre_post*scale(AIMS_total)|ResponseId) + (1|stim),control=lmerControl(optimizer="bobyqa"))

summary(m44)

m5 <- lmer(data=liking_master, scale(value) ~ pre_post*scale(eBMRQ_total) + (pre_post*scale(eBMRQ_total)|ResponseId) + (1|stim),control=lmerControl(optimizer="bobyqa"))
summary(m5)



#did musicality ratings differ by AIMS scores?
m5 <- lmer(data=speech_master, scale(value) ~ stim_type*scale(AIMS_total) + (stim_type*scale(AIMS_total)|ResponseId) + (1|stim),control=lmerControl(optimizer="bobyqa"))
summary(m5)

m6 <- lmer(data=speech_master, scale(value) ~ pre_post*scale(AIMS_total) + (pre_post*scale(AIMS_total)|ResponseId) + (1|stim),control=lmerControl(optimizer="bobyqa"))
summary(m6)

ggplot(data=speech_master, aes(x=AIMS_total, y=value))+
  #geom_point()+
  stat_summary(fun.data = 'mean_se', geom = 'errorbar', width = 0, alpha=0.5) +
  stat_summary(fun.data = 'mean_se', geom = 'pointrange', alpha=0.5)+
  theme_classic()+
  geom_smooth(method="lm") +
  facet_grid(~stim_type) +
  xlab("Total AIMS Score") +
  ylab("Musicality Rating")  
  

#did liking ratings differ by AIMS scores?
m6 <- lmer(data=liking_master, scale(value) ~ stim_type*scale(AIMS_total) + (stim_type*scale(AIMS_total)|ResponseId) + (1|stim),control=lmerControl(optimizer="bobyqa"))
summary(m6)


#change
m10 <- lmer(data=master_prepost, scale(liking_change) ~ scale(musicality_change)*stim_type + (scale(musicality_change)*stim_type|ResponseId) + (1|stim),control=lmerControl(optimizer="bobyqa"))
summary(m10)

#check change correlation 
cor.test(master_long$change, master_long$change)

#check "illusion strength" (average effect of control vs average effect of illusion) (check ani paper w/ adam)

master$BMRQ_total <- master$eBMRQ_total - master$BMRQ_absorption

 

boston_fullgmsi <- read.csv(file='../data/master_bostonfullgmsi.csv')

master_single <- boston_fullgmsi %>%
  distinct(ResponseId,.keep_all = TRUE)

write.csv(master_single, '../data/master_single_boston.csv')


sd(master_single$AIMS_total, na.rm=TRUE) 
  
  
ggplot(data = master_single, aes(x=Gold_musicaltraining, y=AIMS_total)) +
  geom_point()+
  theme_classic() +
  xlab("Total GoldMSI Musical Training Scores")+
  ylab("Total AIMS Scores")

matched_a <- matched %>%
  distinct(ResponseId, .keep_all = T) %>%
  subset(group=="Musical Anhedonics")
  
table(matched_a$Q409)

matched_c <- matched %>%
  distinct(ResponseId, .keep_all = T) %>%
  subset(group=="Matched Controls")

table(matched_c$Q409)

t.test(count(matched_a$Q409), count(matched_c$Q409),paired = FALSE)

test <- lmer(data=matched_c, Q409 ~ group)
```
